Add aggregate API resource
