+++ exercise/cache/content.py
@@ -50,6 +50,7 @@ class CachedContent(ContentMixin, CachedAbstract):
                     'order': o.order,
                     'status': o.status,
                     'name': str(o),
+                    'number': module['number'] + '.' + o.number(),
                     'link': o.get_display_url(),
                     'submittable': False,
                     'submissions_link': o.get_submission_list_url(),
@@ -91,6 +92,7 @@ class CachedContent(ContentMixin, CachedAbstract):
                 'order': module.order,
                 'status': module.status,
                 'name': str(module),
+                'number': str(module.order),
                 'introduction': module.introduction,
                 'link': module.get_absolute_url(),
                 'requirements': [str(r) for r in module.requirements.all()],
